Allow to modify the server timing in runtime.
Motivation:

I have some acceptance tests where I want to tests that if the server
take too long to respond, a message of "Loading more data" is displayed.
Making the interceptor respond without delay in tests is a reasonable
default but for my use case made some things untestable.

What I wanted was to allow in specific tests to set `server.timing = 1000`
and then restore. That was not possible because the route handlers were
registered with the timing at the moment when the configuration is
loaded, by passint that number directly:

```js
  this.stub = function(verb, path, handler, code) {
    var timing = environment === 'test' ? 0 : this.timing;
    /* some code */
    interceptor[verb].call(interceptor, this.namespace + '/' + path, function(request) {
     /* route handling */
    }, timing});
  };
```

It turns out that Pretender is impressively good designed and the
`async` parameter can also be a function that returns the number of
milliseconds.

Thanks to that, the changes are minimal. The only tricky part was how
to test that. I created an acceptance test for that because I cound't
figure out how to test that in isolation.

Check it and tell me what you think.