[closes #50] Disable Mirage if `/server/proxies` present.
In addition to checking if the `/server/proxies` dir exists,
this commit fixes a bug. The previous logic here was wrong.
If proxies are present (either via the `--proxy` arg or the
existence of `/server/proxies`), the user will
not want to use Mirage in development. However, they will
still want to use Mirage for tests, so we need to include
the files in index.js; we simply skip the initializer in
development in those cases.

So, we need to include the files regardless of the presence
of proxies; we only exclude them in production (unless the
force option is present).