Import mirage
The extraction of Mirage from ember-cli-mirage has been a stable
companinion on top of which we have been building production code for
almost 8 months. We've been holding off publishing it as an NPM
package in the hope that it will get merged upstream, or that we can
get the upstream developers on board with supporting this extraction.

At this point though, it looks like this is not very likely to happen
any time soon, and so it's time to move it into bigtest, take
ownership of the maintenance of this fork, and release it to NPM so
that users can install it from the public NMP registry instead of from
git.

This adds what was formerly `mirage-server` into the `packages/mirage`
directory as another bigtest package. The history has been preserved
from both mirage-server and ember-cli-mirage using the tomono.sh
script.