Fix @bigtest/interaction karma webpack config
When a bigtest module is not linked, webpack will not transform any
code from that module within node_modules. This commit is a
quick-fix. The long-term fix would be to build and distribute
es-module ready files from other bigtest packages instead of
referencing `src` in the package.json modules field