Add full support for returning promises in .do() blocks
This was previously, loosely, supported due to the currying nature of
.do() blocks within a promise. This makes stats for these blocks
actually report the true times and return values of the promises.

If a promise takes longer than the remaining timeout, an error will be
thrown after the promise resolves. This same error will now also be
thrown if, at any point in the stack, the elapsed time exceeds the
allotted timeout.