Add docs commands to the remaining packages
This is so the `each-pkg yarn docs` command does not fail when a
package does not have a `docs` script defined in package.json.

The docs script was modified slightly to accept input and output as
arguments. This was required because mirage has a different src
directory than the rest of our packages.