♻Make Client a good ol' simple class
The original implementation had a bunch of complexity to ensure that
operations would be bound to the instance of the class. This just uses
normal old class syntax. The only drawback is that would not be able
to peel off a method and use it elsewhere:

```ts
let { query } = client;
yield query("{ agents { id }}"); // will not work!
```

But if that becomes a requirement, we can use something
like (autoBind)[1]

[1]: https://github.com/sindresorhus/auto-bind