Thinner abstraction on top of fetch
We currently have a test helper method called `get` which abstracts over the fetch API in some ways, but as we make more complicated requests we end up having to replicate most of the options to fetch anyway. So let's just use fetch directly by adding a thin effection API on top of it.