Prepack for CI and for development
Yarn workspaces gives preference to local packages inside the monorepo
over anything listed in the package.json. So, for example, the
contents of the `packages/agent` directory is preferred over the
`@bigest/agent` npm packages.

As we add many independent NPM packages to the monorepo, this makes
development awkward since you need to build your package ahead of
time. This adds a `yarn prepack` script to the root package that fans
out a prepack over all of the packages. That way, you can run it once
before hacking on your individual package, and all the other packages
are ready to be used.

Why `prepack` and not `build`?
------------------------------

prepack is the npm package state that is "just before the
tarball is created". In other words, the exact package contents except
in expanded directory form. This is the level of readiness that we should
expect all of our packages to be in before we start depending on them
from other packages. Otherwise, we may be developing against packages
in a different form than how they'll actually take in production.

The website is the outlier here, but since we might at some point
support deploying it as an NPM package, I went ahead and changed its
build script to be named 'prepack'.

Open Questions
--------------

- How do we override local package directories with packages from NPM?
- How do we hack on local packages and have them rebuild
- automatically.
-