Better and more consistent options
Instead of the Orchestrator having to be supplied with a whole bunch of directories that the caller doesn't care about, let's just give a singular cacheDir, and the orchestrator can decide on a subdirectory structure on its own.

This also makes the options to the subcomponents by being more distinct about the difference between a full path and a directory.