Move to new home in the bigtest monorepo
Now that there is a new home for all of the bigtest packages, server
is the last hold out. It's time to integrate it with the rest of the
dev environment.

This moves all of the server artifacts to their new home in the
`packages/server` directory of `thefrontside/bigtest` repo. It was
mostly left completely in-tact, but a few files were deleted including

* `.circleci` since the new repo uses GH actions for CI
* `.github` directory is deleted, since it is handled by the root of
  the app