Disable convergence tests
Motivation
----------

The @bigtest/convergence tests are flaky and often fail, and block the
code that will ultimately replace it. For example [this failed test
run][1] blocked a completely unrelated PR.

Approach
--------

Since we are going to completely re-write convergence, and the current
test suite is not maintained, running them on every build is dead
weight, especially when they actually block active changes.

[1]: https://github.com/thefrontside/bigtest/pull/194/checks?check_run_id=488143393#step:6:215