Optimize agent for orchestrator-controlled result collection
When we began to implement the result collection for tests in the
orchestrator, the most optimal way to do it was to immediately create
a tree of "result listeners" for each test, step and result in the
entire hiearchy. Because of this, events like 'test:running' are
purely informational, and events like 'test:done' are completely
unecessary.

This PR removes `test:done` entirely, and separates out the
`run:begin` and `run:end` to fire a `lane:begin` and `lane:end` event
from the harness. These are purely informational and not needed except
for perhaps some debugging purposes.