Replace suspend with effection resources
Effection now supports resource objects on an experimental branch, this makes it possible to tie the lifetime of a context to an object, and return it out of its context to pass ownership of the context to the parent. This provides an alternate escape hatch to using suspend, which is more disciplined and composes better.

See: https://github.com/thefrontside/effection.js/pull/89