♻Extract Agent library into re-useable library
We want to write big tests for the bigtest orchestrator. That means
only interfacting with it in the same way as external process will be
interfacing with it.

That mean, only using the graphql API to query and initiate state
changes and to only use external processes communicating over
wesockets to represent the agent.

This refactors the agent code to extract the actual portion concerned
with communicating with the orchestrator into an implementation that
can be shared between the web agent, and any other javascript
environment.

The protocol of allowable messages is implemented using a collection
of TypeScript interface.