Simplify tsconfig.json setup for the entire project (#231)
## Motivation

I'm working on refactoring cli-mob branch to use Atom. In the process, I came across the config.dist.json pattern that we've been using. I found it thoroughly confusing because we have a tsconfig.json that includes tests and a special tsconfig for building package. This is a double negative equivalent of implementation. After some experimentation, it turned out that we don't need to include information about tests in tsconfig.json because `-r ts-node/register` takes care of whats necessary for tests. 

## Approach

Removed all test specific changes in tsconfg.json and deleted all of the tsconfig.dist.json files. 

## TODO

- Move these configs to their own package and consume them from that package