Only generate schema artifacts explicitly
By default, nexus _always_ tries to generate schema artifacts like the
`.graphql` schemafile and the typings. However, [this fails in
production][1] because it tries to load typescript modules which have
long since been compiled into javascript and are not shipped with the
distribution.

Nexus provides a flag to prevent schema generation presumably for
cases precisely such as this one. We use this flag here to stop
generation unless the `BIGTEST_GENERATE_SCHEMA` environment variable
is explicitly set which is done only from the `yarn generate:schema`
step.

While this feels a bit awkward and hackish, it's the only workaround
that I've been able to find.

[1]: https://github.com/thefrontside/bigtest/issues/244