Upgrade effection, improve linkability
One of the challenges of working with effection inside the bigtest
repo has been trying to use a linked version or a preview version and
not have things go terribly awry. This is because each package has it
listed as a runtime dependency. However, most packages don't actually
care that there be a particular version of effection installed, merely
that _somme_ version is available. [I _think_ that this is why you want
to have peer dependencies.][1]

Throughout our tree, we want to make sure that everyone is handed the
same copy of effection (in this case 0.6.2) and the only way to
guarantee that is via a peer dependency.

Now, every library will get the same version of effection that is
loaded by the root level require.

[1: ]https://classic.yarnpkg.com/blog/2018/04/18/dependencies-done-right/