Always handle proxy events within effection context.
It turns out we were cheating a bit and installing our top level event
handlers directly on the proxy `http.Server` object, and not actually
entering an effection context at all until we'd received the upstream
response from the proxy target. This becomes a problem when you are
shutting down your process while you have proxy requests in-flight
since the upstream target will stop responding, causing an uncaught
http exception when the request tries to continue reading bytes from
it.

This converts both of the raw event handlers (web, and websocket) into
`@effection/events` subscriptions that shutdown each request if the
effection context happens to be destroyed mid-request.

In the case of normal http requests we wait until the response is
fully written before destroying the request. In the case of a
websocket, we wait until the `close` event on the websocket since the
socket will, in most cases outlive the request that brought it into
existence. However, in the event that the effection context is halted
_during a websocket upgrade_, then we'll need to destroy the original
request.