Make interactor generic
There are some tricky variance issues with making Interactor generic, we found that using `T extends Interactor<any>` for `find` was the most successful. Presumably (though I haven't verified this) the `any` type is bivariant towards generic types (that is it allows both subtypes AND supertypes to be assigned), so regardless of the inferred variance of the generic type it will always match.

Co-authored-by: Paul Cowan <paul.cowan@cutting.scot>