Prepare the way for node drivers
Motivation
----------

In order to use bigtest on a project that has both a web application
_and_ a server side application, we need the ability to use bigtest to
to test _both_. This means having a node agent that can run node
tests. However, a pre-requisite for having a node agent, is allowing
bigtest to manage node processes in the same way that it manages
browser processes. Currently, the way this works is to abstract away
the process management of agent into a _driver_ which is responsible
for starting the agent process, connecting the agent process
to the orchestrator, and ultimately shutting down the agent when it is
no longer needed. But the only drivers that we have now are web
drivers that manage browser processes.

Approach
--------

This clears the way to make drivers more of a general purpose concept
in BigTest rather than being implicitly local webdriver
instances. Now, the server does not actually reference
`@bigtest/webdriver` at all directly, instead using only the abstract
driver interfaces.

To accomplish this, we adopt a "plugin-like" architecture, where part
of the configuration of the driver is the node module where it can be
found. Whereas before the drivers section the project config looked
like:

```json
{
  "chrome": {
    browserName: "chrome",
    headless: true
  }
}
```

We now specify the module implementing the driver, and a set of
"options" that will be passed through transparently to the driver. So
the previous example will let us say:

```json
{
  "chrome": {
    module: "@bigtest/webdriver",
    options: {
      browserName: "chrome",
      headless: true
    }
  }
}
```

This will allow us to create a node driver that can be configured
with:

```json
{
  "node": {
    module: "@bigtest/nodedriver",
    options: {
      version: 12
      unhandledPromiseRejection: 'strict'
    }
  }
}
```

To implement the driver, the module specified should have a `create`
export that implements the `DriverFactory` interface.