Add globals package for shared global variables
We want bigtest components to be able to share configuration with each
other, without having to explicitly call into each other, and with zero
configuration. The only way to enable this in today's JavaScript is
through shared global variables. To make this a bit nicer and safer, we
add a package which can be used to access these variables in a typesafe
and convenient way.

This paves the way for #315 and also for completing #286.