Ensure app is returning html before considered ready
It isn't enough to know that a server is accepting connections when
considering an app to be ready because there are cases (like a
create-react-app devserver) where the server will be bound to a port,
but not yet providing responses. In the case of create-react-app, it
won't actually respond until the app is first built.

Now that we have `@effection/fetch`, this uses it to check that we're
actually getting a response before considering the app "reachable".