inject `process.env.NODE_ENV` into environment
Many third party libraries use `process.env.NODE_ENV` even when
running on the browser. While this isn't strictly our responsibility
to account for this, there's little we can do about it. In the case
that fomented this change, we had a package (@mirage/graphql) that
dependend on a package(graphql) that referenced
`process.env.NODE_ENV`. Parcel and webpack both make this variable
available, and so we should too.

In order to define the global `process.env` variable, we're using what
appears to be the most well-worn rollup plugin to do so. However,
because of the way its exports are constructed, it did require us to
tweak our `tsconfig` to use `esModuleInterop`.

I hardcoded the `NODE_ENV` to 'production' since there's no reason to
slow down elements of the test bundle itself. We may want to consider
allowing to set this to some other value at some point, but not too
sure what form that would take.

Finally, I added a preference for ES modules as an entry point since they
are usually better behaved with TypeScript.