Consume websockets on the server using @effection/subscription
In order to write tests for the bigtest agent, we need to simulate an
orchestrator agent handler that the agent can connect to. However,
the @bigtest/effection-express package requires you to write your
websocket handler with the deprecated mailbox handler interface.

Rather than replace the mailbox interface, this adds the subscription
interface along side. That way, consumers of effection-express can
migrate over gradually over time.

In order to verify both interfaces, tests have been wrapped around the
websocket API for both the Mailbox interface as well as the
subscription interface.