Prevent build error spam
Closes #532

This is a quick fix which ensures we only print the bundler error message when the state has actually changed, preventing the console from being spammed to death.

Additionally it only prints the error message, and not the stack trace, since that trace is completely useless.

While we will want a more elegant solution for this eventually, it'll do fine for now.