TextField interactor matches wider range of inputs
Currently out TextField interactor is pretty strict, it only matches
`[type=text]` inputs, and those that do not have a type attribute at
all. With this change it will also match all of the text-like fields,
like `[type=email]` or `[type=password]`, it will also match an input
with a completely made up type. This behaviour is more inline with how
browsers display these fields.

Note that we do *not* match any of the date related fields, this is
because those fields differ in their functioning across browsers, and
simply entering text into them is unlikely to have the desired result.
This is actually a major pain point when using Selenium, so we should
avoid this problem.