Catch unexpected exceptions and offer to report them as bugs
It's a tough look whenever something terrible happens and the server
or the CLI halts without doing the thing that you were asking it to
do. For example, we were having an issue with our proxy server
crashing on canceled requests with:

```
Error: unexpected end of file
    at Zlib.zlibOnError [as onerror] (zlib.js:180:17) {
  errno: -5,
  code: 'Z_BUF_ERROR'
}
```

It's even worse when that happens, and it silently makes a terrible
impression on our users and we never even have the notification that
it happened.

This change catches all top-level exceptions that are _not_ effetion
`MainError`s and prints out an explanation and generates a link to a
new github issue with several diagnostic  values pre-filled in.