Do not wrap action functions in a convergence
Closes #691

This removes the automatically inserted convergence for action
functions, which means that actions do not converge automatically
anymore. Instead the user will need to ensure they converge by awaiting
a converging function inside the action. This could be either an
action/assertion on another interactor, or it could be one of the built
in converging functions on the interactor itself. Most likely, the user
will want to call `perform`. To remove some boilerplate for this, we
have the `perform` function, which most actions already use, so this
will not be a breaking change for most interactors.

This also means that "normal" actions which use `perform` to not
nest calls of converge. Additionally, this change explicitly awaits the
converge call in the `perform` helper function, removing the useless
wrapping in an interaction as well.

I've been wracking my brain trying to come up with a way of testing
this, but any test for this change would really be testing on the
absence of something, e.g. testing that the async function is *not*
exectuted multiple times, which seems like a bit of a silly test to me.