Add `changeOrigin` flag to application proxy server
Whenever we fetch from an https proxy, by default the origin of the
request to the target will be made with the host of the proxy
server. However, when verifying a public url over TLS it will see that
the certificate does not have `localhost` as one of its altnames and
so will fail. In other words a cert for server.frontside.com will
validate content from neither google.com nor localhost... only
server.frontside.com which causes a validation failure.

This sets the `changeOrigin` flag on the proxy server which causses
the outgoing requests from the proxy server [to have the same origin as
the target][1]. Iin the case above, would cause the proxy request to
server.frontside.com to have that origin as well, so that TLS
validation succeeds.

It's difficult to say what other impact this change will have, which
is the only thing that worries me at the moment.

> resolves https://github.com/thefrontside/bigtest/issues/781

[1]: https://github.com/http-party/node-http-proxy/blob/9b96cd725127a024dabebec6c7ea8c807272223d/test/lib-http-proxy-common-test.js#L291-L315